Rename alias module to customcommands
