+++ src/main/java/uk/firedev/firefly/modules/customcommands/CommandBuilder.java
@@ -1,4 +1,4 @@
-package uk.firedev.firefly.modules.alias;
+package uk.firedev.firefly.modules.customcommands;
 
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
